feat: channel fixes
