+++ src/components/app/guild/side-bar.tsx
@@ -26,7 +26,6 @@ import {
   readyStore,
 } from "@/utils/stores";
 import { useRecoilState } from "recoil";
-import NextLink from "next/link";
 import { sortChannels } from "@/utils/sortChannels";
 import GuildSettings from "@/components/app/guild/settings";
 import GuildInvites from "@/components/app/guild/invites";
@@ -35,8 +34,8 @@ import GuildMessageContainer from "@/components/app/guild/messageContainer.tsx";
 import { IoPeople } from "react-icons/io5";
 import CreateChannel from "@/components/app/guild/createChannel.tsx";
 import BaseChannel from "$/Client/Structures/Channels/BaseChannel.ts";
-import constants from "$/utils/constants.ts";
 import getGuildName from "@/utils/getGuildName.ts";
+import Channel from "./channels/index.tsx";
 
 const GuildSideBar = ({ children, noMemberBar, noTextBox, noChannelTopic }: { children?: ReactNode, noMemberBar?: boolean; noTextBox?: boolean; noChannelTopic?: boolean; }) => {
   const [client] = useRecoilState(clientStore);
@@ -63,14 +62,6 @@ const GuildSideBar = ({ children, noMemberBar, noTextBox, noChannelTopic }: { ch
     // onOpen: createChannelOnOpen,
     onClose: createChannelOnClose,
   } = useDisclosure();
-
-  const getChannelName = (name: string) => {
-    if (name.length > 18) {
-      return name.slice(0, 18) + "...";
-    } else {
-      return name;
-    }
-  };
 
   useEffect(() => {
     if (client.currentGuild) {
@@ -163,112 +154,10 @@ const GuildSideBar = ({ children, noMemberBar, noTextBox, noChannelTopic }: { ch
             aria-label="Main Navigation"
           >
             <div>
-              {sortedChannelGroups
-                ? sortedChannelGroups.map((channel) => {
-                  if (collapsedChannelsList.includes(channel.parentId ?? ""))
-                    return null;
-                  return (
-                    <Flex
-                      py={1}
-                      ml={5}
-                      key={channel?.id}
-                      bg={
-                        channel?.id === client.currentChannel?.id
-                          ? "brand.500"
-                          : ""
-                      }
-                    >
-                      <div>
-                        <Flex>
-                          {client.currentChannel?.type === constants.channelTypes.GuildText && (
-                            <svg
-                              width="20px"
-                              height="20px"
-                              viewBox="0 0 0.72 0.72"
-                              fill="none"
-                              xmlns="http://www.w3.org/2000/svg"
-                            >
-                              <path
-                                d="m0.21 0.57 0.12 -0.42m0.06 0.42 0.12 -0.42m0.06 0.12H0.195m0.33 0.18H0.15"
-                                stroke="#ffff"
-                                strokeWidth="0.06"
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                              />
-                            </svg>
-                          )}
-
-                          {client.currentChannel?.isVoiceBased() && (
-                            <svg className="w-4 h-4 mr-2" viewBox="0 0 24 24">
-                              <path
-                                fill="currentColor"
-                                d="M12 2c-4.4 0-8 3.6-8 8v4c0 4.4 3.6 8 8 8s8-3.6 8-8v-4c0-4.4-3.6-8-8-8zm6 12c0 3.3-2.7 6-6 6s-6-2.7-6-6v-4c0-3.3 2.7-6 6-6s6 2.7 6 6v4zm-4-4h-2v-4h2v4zm-4 0h-2v-4h2v4zm-3 4h-2v-4h2v4zm-1-6h-2v-4h2v4z"
-                              />
-                            </svg>
-                          )}
-
-                          {channel?.type === constants.channelTypes.GuildCategory && (
-                            <Flex>
-                              <Text color={"white"}>
-                                {getChannelName(channel.name)}
-                              </Text>
-                              <svg
-                                width="16px"
-                                height="16px"
-                                viewBox="0 0 0.72 0.72"
-                                fill="none"
-                                xmlns="http://www.w3.org/2000/svg"
-                                style={{
-                                  transform: collapsedChannelsList.includes(
-                                    channel.id,
-                                  )
-                                    ? "rotate(270deg);"
-                                    : "",
-                                }}
-                                className="transform rotate-270 transition-transform duration-250"
-                              >
-                                <path
-                                  fillRule="evenodd"
-                                  clipRule="evenodd"
-                                  d="M0.381 0.441a0.03 0.03 0 0 1 -0.042 0l-0.15 -0.15a0.03 0.03 0 0 1 0.042 -0.042L0.36 0.378l0.129 -0.129a0.03 0.03 0 1 1 0.042 0.042l-0.15 0.15Z"
-                                  fill="#fff"
-                                />
-                              </svg>
-                            </Flex>
-                          )}
-
-                          {client.currentChannel?.type === constants.channelTypes.GuildNews && (
-                            <svg
-                              xmlns="http://www.w3.org/2000/svg"
-                              width="18px"
-                              height="18px"
-                              viewBox="0 0 24 24"
-                              fill="none"
-                              stroke="#fff"
-                              strokeWidth="2"
-                              strokeLinecap="round"
-                              strokeLinejoin="round"
-                            >
-                              <path d="M3 11l18-5v12L3 14v-3z" />
-                              <path d="M11.6 16.8a3 3 0 11-5.8-1.6" />
-                            </svg>
-                          )}
-
-                          {channel?.type !== constants.channelTypes.GuildCategory && (
-                            <NextLink
-                              href={`/app/guilds/${channel?.guildId}/channels/${channel?.id}`}
-                            >
-                              <Text color={"white"}>
-                                {getChannelName(channel?.name)}
-                              </Text>
-                            </NextLink>
-                          )}
-                        </Flex>
-                      </div>
-                    </Flex>
-                  );
-                })
-                : null}
+              {sortedChannelGroups.map((channel) => {
+                  if (collapsedChannelsList.includes(channel.parentId ?? "")) return null;
+                  return <Channel channel={channel} key={channel.id} />;
+                })}
             </div>
           </Flex>
         </Box>
@@ -278,7 +167,7 @@ const GuildSideBar = ({ children, noMemberBar, noTextBox, noChannelTopic }: { ch
         <Box flex="1" justifyContent="center" maxWidth="calc(100% - 400px)">
           {!noChannelTopic && (
             <Box pos="sticky" top={0} zIndex={10} bg={background} p={2}>
-              <Text>{`#${client.currentChannel?.name}`}</Text>
+              <Text>{`# ${client.currentChannel?.name}`}</Text>
             </Box>
           )}
